+++ app/src/main/java/com/lmizuno/smallnotesmanager/DAO/ItemDao.kt
@@ -15,10 +15,10 @@ interface ItemDao {
     fun insert(item: Item): Long
 
     @Query("SELECT * FROM items WHERE itemId = :itemId")
-    fun getItemById(itemId: Int): Item?
+    fun getItemById(itemId: Long): Item?
 
     @Query("SELECT * FROM items WHERE collectionId = :collectionId ORDER BY itemId ASC")
-    fun getItemsByCollection(collectionId: Int): List<Item>
+    fun getItemsByCollection(collectionId: Long): List<Item>
 
     @Update(onConflict = OnConflictStrategy.ABORT)
     fun update(item: Item): Int
