+++ app/src/main/java/com/lmizuno/smallnotesmanager/Models/Collection.kt
@@ -1,4 +1,5 @@
 package com.lmizuno.smallnotesmanager.Models
+
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import java.io.Serializable
@@ -6,7 +7,7 @@ import java.io.Serializable
 @Entity(tableName = "collections")
 data class Collection(
     @PrimaryKey(autoGenerate = true)
-    var collectionId: Int = 0,
+    var collectionId: Long = 0,
     var name: String,
     var description: String
-):Serializable
+) : Serializable
