+++ app/src/main/java/com/lmizuno/smallnotesmanager/Models/Item.kt
@@ -6,18 +6,20 @@ import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
 import java.io.Serializable
 
-@Entity(tableName = "items",
+@Entity(
+    tableName = "items",
     foreignKeys = [ForeignKey(
         entity = Collection::class,
         parentColumns = ["collectionId"],
         childColumns = ["collectionId"],
         onDelete = ForeignKey.CASCADE
-    )])
+    )]
+)
 data class Item(
     @PrimaryKey(autoGenerate = true)
-    var itemId: Int,
+    var itemId: Long,
     @ColumnInfo(index = true)
-    var collectionId: Int,
+    var collectionId: Long,
     var title: String,
     var content: String
-): Serializable
+) : Serializable
